requirements.txt
cGraph==0.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
deap==1.3.3
Deprecated==1.2.13
fonttools==4.38.0
graphviz==0.20.1
kiwisolver==1.4.4
matplotlib==3.6.2
networkx==2.8.8
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
pygraphviz==1.10
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tabulate==0.8.10
tsplib95==0.7.1
wrapt==1.14.1
